+++ csatar-plugins/forms/lang/en/lang.php
@@ -49,10 +49,11 @@
                     'recordKeyNotSelected'  => "Record key parameter is not specified."
                 ],
             ],
-            'select' => 'Select',
-            '2FAtoRead' => 'You need to authenticate with 2FA in order to see the following information: ',
+            'select'      => 'Select',
+            '2FAtoRead'   => 'You need to authenticate with 2FA in order to see the following information: ',
             '2FAtoModify' => 'You need to authenticate with 2FA in order to fill the following information: ',
             '2FAtoModify' => 'You need to authenticate with 2FA in order to modify the following information: ',
+            '2FANeeded'   => 'You need to authenticate with 2FA in order to perform this action.',
         ],
     ],
     'widgets' => [