+++ csatar-plugins/csatar/lang/en/lang.php
@@ -572,6 +572,8 @@ return [
                 'membershipCards' => 'Membership Cards',
                 'issued_date_time' => 'Issued',
                 'active' => 'Active',
+                'requests' => 'Requests',
+                'exportRequests' => 'Export requests to .csv',
             ],
         ],