+++ csatar-plugins/csatar/lang/en/lang.php
@@ -540,6 +540,7 @@
                 'endDate' => 'End date',
                 'breadcrumb' => 'Mandate types',
                 'activeMandateDeleteError' => 'There exist active Mandates of %name type, thus this Mandate type cannot be deleted.',
+                'isHiddenFrontend' => 'Hidden on frontend',
             ],
             'mandate' => [
                 'mandate' => 'Mandate',