+++ Changelog.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- CS-458 - Added buttons partial, buttons partial can handle passed permission value
+- CS-458 - Changed buttons on every component to one of the buttons partial, expect gallery
 - CS-412 - The form on mobile view is now displayed on a single column, instead of two
 - CS-412 - Fixed overflow problem for dropdown boxes on mobile view
 - CS-148 - Updated Terms and conditions page