+++ Changelog.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- CS-466 - Created migrations for "Knowledge repository" models
 - CS-458 - Added buttons partial, buttons partial can handle passed permission value
 - CS-458 - Changed buttons on every component to one of the buttons partial, expect gallery
 - CS-412 - The form on mobile view is now displayed on a single column, instead of two