+++ Changelog.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- CS-451 - Membership card cannot be activated if the assigned scout is inactive
 - CS-462 - Added warning to team report if any scout doesn't have registration form uploaded
 - CS-436 - Added option to hide mandate types on frontend
 - CS-466 - Created migrations for "Knowledge repository" models