+++ Changelog.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+## 1.8.0
+### 2023-02-06
+
 - CS-458 - Added buttons partial, buttons partial can handle passed permission value
 - CS-458 - Changed buttons on every component to one of the buttons partial, expect gallery
 - CS-444 - Personalized TAB design and added to the design system